core: Add utility function to set object properties
Add a new cog_apply_properties_from_key_file() function which iterates
over writable properties of an object and sets their values from the
items of a GKeyFile group with the same names as the properties. The
goal is to use this to ease picking configuration options for the
different platform modules--as long as they define their settings as
properties of some object.